+++ tests/consumers/esm.js
@@ -1,0 +1,4 @@
+import { alert } from 'forgedialog/alert';
+import { makeDraggable } from 'forgedialog/interactions';
+
+export { alert, makeDraggable };